Refactor works but not ideal
