+++ src/webparts/v2NpmBanner/components/PropPaneCols/OnClickHelpers.ts
@@ -1,4 +1,4 @@
-import { IMinField } from "./PropPaneColsClass";
+import { IMinField } from "./components/IPropPaneColsProps";
 
 
 /**
